refactor: 상품의 남은 재고 계산, 상품 가격 정보 포멧팅 함수 개선
- formatPrice 함수는 어드민 여부에 따라 분기처리만 하고, formatPriceWithSoldOut으로 솔드아웃 처리 추가
- getRemainingStock 함수에서 사용되는 cart 배열을 인자로 받아서 순수함수로 변경
