+++ src/basic/App.tsx
@@ -9,6 +9,8 @@ import { initialCoupons } from './data/mockCoupons';
 import { useProducts } from './hooks/product/useProducts';
 import { useNotifications } from './hooks/notifications/useNotifications';
 import { useProductForm } from './hooks/product/useProductForm';
+import { getRemainingStock } from './utils/calculations/stockCalculations';
+import { formatPriceWithSoldOut } from './utils/formatters/priceFormatter';
 
 const App = () => {
   const { products, deleteProduct, updateProduct, addProduct } = useProducts();
@@ -58,21 +60,6 @@ const App = () => {
     discountType: 'amount' as 'amount' | 'percentage',
     discountValue: 0,
   });
-
-  const formatPrice = (price: number, productId?: string): string => {
-    if (productId) {
-      const product = products.find((p) => p.id === productId);
-      if (product && getRemainingStock(product) <= 0) {
-        return 'SOLD OUT';
-      }
-    }
-
-    if (isAdmin) {
-      return `${price.toLocaleString()}원`;
-    }
-
-    return `₩${price.toLocaleString()}`;
-  };
 
   const getMaxApplicableDiscount = (item: CartItem): number => {
     const { discounts } = item.product;
@@ -129,13 +116,6 @@ const App = () => {
     };
   };
 
-  const getRemainingStock = (product: Product): number => {
-    const cartItem = cart.find((item) => item.product.id === product.id);
-    const remaining = product.stock - (cartItem?.quantity || 0);
-
-    return remaining;
-  };
-
   const [totalCartItem, setTotalCartItem] = useState(0);
 
   useEffect(() => {
@@ -164,7 +144,7 @@ const App = () => {
 
   const addToCart = useCallback(
     (product: ProductWithUI) => {
-      const remainingStock = getRemainingStock(product);
+      const remainingStock = getRemainingStock(product, cart);
       if (remainingStock <= 0) {
         addNotification('재고가 부족합니다!', 'error');
         return;
@@ -400,7 +380,11 @@ const App = () => {
                             {product.name}
                           </td>
                           <td className='px-6 py-4 whitespace-nowrap text-sm text-gray-500'>
-                            {formatPrice(product.price, product.id)}
+                            {formatPriceWithSoldOut(
+                              product.price,
+                              isAdmin,
+                              getRemainingStock(product, cart) <= 0,
+                            )}
                           </td>
                           <td className='px-6 py-4 whitespace-nowrap text-sm text-gray-500'>
                             <span
@@ -860,7 +844,7 @@ const App = () => {
                 ) : (
                   <div className='grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-4'>
                     {filteredProducts.map((product) => {
-                      const remainingStock = getRemainingStock(product);
+                      const remainingStock = getRemainingStock(product, cart);
 
                       return (
                         <div
@@ -908,7 +892,11 @@ const App = () => {
                             {/* 가격 정보 */}
                             <div className='mb-3'>
                               <p className='text-lg font-bold text-gray-900'>
-                                {formatPrice(product.price, product.id)}
+                                {formatPriceWithSoldOut(
+                                  product.price,
+                                  isAdmin,
+                                  getRemainingStock(product, cart) <= 0,
+                                )}
                               </p>
                               {product.discounts.length > 0 && (
                                 <p className='text-xs text-gray-500'>
