+++ src/basic/utils/formatters/priceFormatter.ts
@@ -1,0 +1,18 @@
+export const formatPriceWithSoldOut = (
+  price: number,
+  isAdmin: boolean,
+  isSoldOut: boolean,
+): string => {
+  if (isSoldOut) {
+    return 'SOLD OUT';
+  }
+  return formatPrice(price, isAdmin);
+};
+
+export const formatPrice = (price: number, isAdmin: boolean): string => {
+  if (isAdmin) {
+    return `${price.toLocaleString()}원`;
+  }
+
+  return `₩${price.toLocaleString()}`;
+};
